requirements-slack.txt
# Slack Integration Dependencies
slack-bolt>=1.18.0
slack-sdk>=3.26.0
flask>=3.0.0
aiohttp>=3.9.0

# Existing dependencies (from pyproject.toml)
python-dotenv>=1.0.0
requests>=2.31.0
prefect>=2.14.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
minio>=7.2.0
pydantic>=2.0.0
